GlobalContact created
Moved DirSearch and GlobalContact related functions to Friendica\Model namespace
