+++ mod/allfriends.php
@@ -1,13 +1,16 @@
 <?php
-
+/**
+ * @file mod/allfriends.php
+ */
 use Friendica\App;
 use Friendica\Core\System;
 use Friendica\Database\DBM;
+use Friendica\Model\GlobalContact;
 
-require_once('include/socgraph.php');
-require_once('include/Contact.php');
-require_once('include/contact_selectors.php');
-require_once('mod/contacts.php');
+require_once 'include/socgraph.php';
+require_once 'include/Contact.php';
+require_once 'include/contact_selectors.php';
+require_once 'mod/contacts.php';
 
 function allfriends_content(App $a) {
 
@@ -39,12 +42,12 @@ function allfriends_content(App $a) {
 	$a->page['aside'] = "";
 	profile_load($a, "", 0, get_contact_details_by_url($c[0]["url"]));
 
-	$total = count_all_friends(local_user(), $cid);
+	$total = GlobalContact::countAllFriends(local_user(), $cid);
 
 	if(count($total))
 		$a->set_pager_total($total);
 
-	$r = all_friends(local_user(), $cid, $a->pager['start'], $a->pager['itemspage']);
+	$r = GlobalContact::allFriends(local_user(), $cid, $a->pager['start'], $a->pager['itemspage']);
 
 	if (! DBM::is_result($r)) {
 		$o .= t('No friends to display.');
